Split watch page signal helpers
