+++ frontend/server/watch-page-signals/normalization.ts
@@ -1,0 +1,26 @@
+import type { RawRecord } from './types';
+
+export function asRecord(value: unknown): RawRecord | null {
+  return value && typeof value === 'object' && !Array.isArray(value) ? (value as RawRecord) : null;
+}
+
+export function asString(value: unknown): string | null {
+  return typeof value === 'string' && value.trim().length ? value.trim() : null;
+}
+
+export function asNumber(value: unknown): number | null {
+  if (typeof value === 'number' && Number.isFinite(value)) return value;
+  if (typeof value === 'string' && value.trim().length) {
+    const parsed = Number(value);
+    return Number.isFinite(parsed) ? parsed : null;
+  }
+  return null;
+}
+
+export function asBoolean(value: unknown): boolean | null {
+  return typeof value === 'boolean' ? value : null;
+}
+
+export function asArray(value: unknown): unknown[] {
+  return Array.isArray(value) ? value : [];
+}
